+++ package.json
@@ -4,6 +4,7 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
+    "test": "vitest",
     "dev": "vite",
     "build": "tsc -b && vite build",
     "types:check": "tsc",
@@ -30,7 +31,8 @@
     "prettier": "3.3.3",
     "tailwindcss": "3.4.15",
     "typescript": "5.6.2",
-    "vite": "5.4.7"
+    "vite": "5.4.7",
+    "vitest": "2.1.6"
   },
   "packageManager": "yarn@4.4.1"
 }
